homework 2 server rendering
