--- index.html
@@ -1,11 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Homework 1</title>
-</head>
-<body>
-    <div id="root"></div>
-    <script src="index.js"></script>
-</body>
-</html>
