+++ index.js
@@ -1,86 +1,24 @@
-import React, { Component } from 'react';
-import ReactDOM from 'react-dom';
-import Products from './constants/Products';
+const express = require('express');
 
-class Image extends Component {
-    constructor(props) {
-        super(props);
-    }
+require('@babel/register');
 
-    render() {
-        const product = this.props.product;
-        return (
-            <img src={product.imageUrl} width='100' height='100' alt={product.title}/>
-        )
-    }
-}
+const render = require('./render').default;
 
-class TextBox extends Component {
-    constructor(props) {
-        super(props);
-    }
+const app = express();
 
-    render() {
-        const product = this.props.product;
-        return (
-            <span>{product.title}</span>
-        )
-    }
-}
+app.get('/', (rew, res) => {
+   const response = `
+        <!DOCTYPE html>
+        <html lang="en">
+            <head>
+                <title>Homework 2, Server rendering</title>
+            </head>
+            <body>
+                <div id="root">${render()}</div>
+            </body>
+        </html>
+   `;
+   res.send(response);
+});
 
-class Price extends Component {
-    constructor(props) {
-        super(props);
-    }
-
-    render(){
-        const product = this.props.product;
-        return <span>{product.price}</span>
-    }
-}
-
-class ProductCard extends Component {
-    constructor(props) {
-        super(props);
-    }
-    render() {
-        const product = this.props.product;
-        return (
-            <div>
-                <TextBox product={product}/>
-                <Price product={product}/>
-                <Image product={product}/>
-            </div>
-        )
-    }
-}
-
-class Catalog extends Component {
-    render() {
-        const { products } = this.props;
-        return (
-            <ul>
-                {
-                    products.map(product => {
-                        return <li key = {product.id}><ProductCard product={product}/></li>
-                    })
-                }
-            </ul>
-        )
-    }
-}
-
-class CatalogPage extends Component {
-    render() {
-        const products = Products;
-        return (
-            <Catalog products={products}/>
-        )
-    }
-
-}
-
-ReactDOM.render(
-    <CatalogPage/>,
-    document.getElementById('root')
-)
+app.listen(3001, () => console.log('Server is listening port 3001'))
